Fix rsync flags.
Previously I had size-only which is obviously wrong because often the
size won't change. What I was trying to do was `-I` which doesn't
attempt to check if the files are the same but copies them all. This is
what we want because the files we are overwriting are probably
different. We also use `-W` to do `whole file` transfer as on a local
system trying to save transfer by reading the existing destination file
is a waste of time.